requirements-etl.txt
pandas==2.0.1
fastparquet==2023.4.0
pyarrow==12.0.0
pandas-gbq==0.19.2
functions-framework==3.*
google-cloud-logging==3.4.0
python-dateutil==2.8.2